requirements.txt
fastapi>=0.115,<0.116
SQLAlchemy>=2.0,<2.1
alembic>=1.16,<1.17
asyncpg>=0.30,<0.31
greenlet>=3.2,<3.3
passlib>=1.7,<1.8
bcrypt>=4.3,<4.4
pyhumps==3.8.0
gunicorn>=23.0,<23.1
uvicorn-worker>=0.3,<0.4